requirements.txt
flask
langchain
langchain-community
langchain-huggingface
python-dotenv
sentence-transformers
chromadb